+++ lib/dry/types/options.rb
@@ -26,6 +26,11 @@ module Dry
       def meta(data = nil)
         data ? with(meta: @meta.merge(data)) : @meta
       end
+
+      # @return [Boolean]
+      def optional?
+        self.class == Sum && left.primitive == NilClass
+      end
     end
   end
 end