+++ spec/dry/types/sum_spec.rb
@@ -11,6 +11,20 @@ RSpec.describe Dry::Types::Sum do
     end
   end
 
+  describe '#optional?' do
+    it 'return true if left side is nil' do
+      type = Dry::Types['strict.nil'] | Dry::Types['string']
+
+      expect(type.optional?).to eq true
+    end
+
+    it 'return false if left is not nil' do
+      type = Dry::Types['string'] | Dry::Types['nil']
+
+      expect(type.optional?).to eq false
+    end
+  end
+
   describe '#[]' do
     it 'works with two pass-through types' do
       type = Dry::Types['int'] | Dry::Types['string']